Move DnsSpec to multi-jvm test, #22330
* run in separate jvm to avoid issues with parallel test execution
  when modifying global System.properties

(cherry picked from commit 1586afe79be568e3815d62d2fb0179a8a017d568)
